Double Declaration of "te"
The variable "te" is already declared as protected in AetherCraftGui on line 16. This declaration here causes the game to crash with a null pointer when you right click on this block, as the requestSync() function uses the protected "te" over the private one initialized and declared through this class. Removing this line allows the block to function properly.