Fall distance is not reset on activation of ring
The fall distance should be reset on each ring activation to prevent short distances from killing the player. One example would be if you were to fall off a 50 block cliff and activate the ring after you fell 49 blocks, if you deactivated the ring again you would still take 50 blocks of fall damage instead of the 1 that it should be.